add aws deploy file and update sam files
Create github workflow file that allows the lambda functions to be automatically deployed when a PR is merged into the main branch of the repo.
In addition, a new event file was created to trigger lambda locally, and samconfig.toml and template.yaml were updated to accurately reflect our
project now that we are past a "hello world" state.